# api-gateway/docs/swagger.yaml
basePath: /api/v1
definitions:
  internal_handler.APIResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: study.khoipn.id.vn
info:
  contact:
    email: support@study.khoipn.id.vn
    name: Study Platform Support
    url: https://study.khoipn.id.vn/support
  description: API Gateway for Study Platform - A comprehensive learning management
    system
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://study.khoipn.id.vn/terms
  title: Study Platform API
  version: "1.0"
paths:
...
  /forum/topics:
    get:
      description: Get list of forum topics with pagination
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: page_size
        type: integer
      - description: Filter by course
        in: query
        name: course_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of topics
          schema:
            $ref: '#/definitions/internal_handler.APIResponse'
      summary: List forum topics
      tags:
      - Forum
    post:
      consumes:
      - application/json
      description: Create new forum topic
      produces:
      - application/json
      responses:
        "201":
          description: Topic created
          schema:
            $ref: '#/definitions/internal_handler.APIResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/internal_handler.APIResponse'
      security:
      - BearerAuth: []
      summary: Create topic
      tags:
      - Forum
  /forum/topics/{topicId}:
    get:
      description: Get forum topic by ID
      parameters:
      - description: Topic ID
        in: path
        name: topicId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Topic details
          schema:
            $ref: '#/definitions/internal_handler.APIResponse'
        "404":
          description: Topic not found
          schema:
            $ref: '#/definitions/internal_handler.APIResponse'
      summary: Get topic
      tags:
      - Forum
...
produces:
- application/json
schemes:
- https
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...